requirements.txt
alembic==1.3.1
astroid==2.3.3
Click==7.0
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lxml==4.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
torch==1.3.1
torchvision==0.4.2
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
